% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ctcorrelator.R
\name{ctcorrelator}
\alias{ctcorrelator}
\title{CT Correlation Finder}
\arguments{
\item{dat_data}{Dataframe to compare. Columns must be labeled "town" and "variable". Highly recommend running ctnamecleaner() on it to make sure town names match.}

\item{p}{Sets Pearson product-moment correlation coefficient. Default is .9.}
}
\value{
Whatever
}
\description{
Compares an array town data to ~500 variables to check if there are any strong correlations. Exports two CSV files and one chart.
}
\examples{
\dontrun{
ctcorrelator(dat_data, p=.9)
}
}

