% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/cpopulator.R
\name{ctpopulator}
\alias{ctpopulator}
\title{CT Population Appender}
\arguments{
\item{name}{Column with town names}

\item{data}{Name of dataframe}
}
\value{
Whatever
}
\description{
Adds population of CT towns in extra column to dataframe.
}
\examples{
\dontrun{
ctpopulator(name_column, ctdata, filename="analysis")
}
}

