% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ctnamecleaner.R
\name{ctnamecleaner}
\alias{ctnamecleaner}
\title{CT Town name cleaner}
\arguments{
\item{name}{Column with town names}

\item{data}{Name of dataframe}

\item{filename}{Name of CSV to save}

\item{case}{Output of town string. Options are \code{Upper}, \code{Lower}, and \code{Title}}
}
\value{
Whatever
}
\description{
Cleans up Connecticut town names and adds population if the user wants and then exports the modified dataframe as a CSV.
}
\examples{
#ctnamecleaner(name_column, ctdata, filename="analysis", case="Upper")
}

