% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpopulator.R, R/ctpopulator.R
\name{ctpopulator}
\alias{ctpopulator}
\title{CT Population Appender}
\arguments{
\item{name}{Column with town names}

\item{data}{Name of dataframe}

\item{name}{Column with town names}

\item{data}{Name of dataframe}

\item{filename}{Name of CSV to export. Default is `nope` which creates dataframe without exporting.}
}
\value{
Whatever

Whatever
}
\description{
Adds population of CT towns in extra column to dataframe.

Adds population of CT towns in extra column to dataframe.
}
\examples{
\dontrun{
ctpopulator(name_column, ctdata, filename="analysis")
}
\dontrun{
ctpopulator(name_column, ctdata, filename="analysis")
}
}

